requirements.txt
# Backend dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
httpx>=0.26.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.66.0
python-dateutil>=2.8.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0

# Dev dependencies
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
